# perf: bypass method validation