<details>
<summary>Benchmark Script</summary>

```js
import { bench, run } from 'mitata'
import { Request } from '../index.js'
// avoid JIT bias
bench('noop', () => {})
bench('noop', () => {})
bench('noop', () => {})
bench('noop', () => {})
bench('noop', () => {})
bench('noop', () => {})

const input = "https://localhost/post"

const method = "post"

bench('new Request(input, { method })', () => {
    return new Request(input, { method })
})

await new Promise((resolve) => setTimeout(resolve, 7000))

await run()

```
</details>

- *main*
```
benchmark                           time (avg)             (min … max)       p75       p99      p995
---------------------------------------------------------------------- -----------------------------
new Request(input, { method })   12.47 µs/iter      (8.5 µs … 2.37 ms)   13.6 µs   27.4 µs   38.4 µs
```

- *this patch*
```
benchmark                           time (avg)             (min … max)       p75       p99      p995
---------------------------------------------------------------------- -----------------------------
new Request(input, { method })   11.79 µs/iter      (8.5 µs … 2.17 ms)   12.1 µs     24 µs     33 µs
```