@mcollina 2024-01-03
> I don't think it's worth it

@KhafraDev why?